+++ src/scene/scenemenu.hpp
@@ -27,7 +27,7 @@ private:
 
 	ce::texture tex_arrow;
 	direction arrow_dir;
-	int current = 0;
+	size_t current = 0;
 
 	ce::font fnt_menu;
 	std::vector<ce::text> texts;
@@ -36,6 +36,5 @@ private:
 
 	auto texts_height() -> int;
 
-	void set_current(int value);
-
+	void set_current(size_t value);
 };
